Aggiunte migliorie e il DbContextPool

+++ src/Models/Entities/Courses.cs
@@ -17,7 +17,7 @@ namespace CorsoDotNet.Models.Entities
             Lessons = new HashSet<Lesson>();
         }
 
-        public long Id { get; private set; }
+        public int Id { get; private set; }
         public string Title { get; private set; }
         public string Description { get; private set; }
         public string ImagePath { get; private set; }
